updates after rebase, moved 'AssertEqualAst' into '/fastbelt/test' enabling re-use
