+++ test/doc_fixture.go
@@ -1,3 +1,7 @@
+// Copyright 2026 TypeFox GmbH
+// This program and the accompanying materials are made available under the
+// terms of the MIT License, which is available in the project root.
+
 package test
 
 import (
